+++ config/src/main/resources/shared/oracle-service.yml
@@ -7,10 +7,6 @@ spring:
       bindings:
         sannya: 
           group: oracle
-  data:
-    elasticsearch:
-      cluster-name: delphi
-    cluster-nodes: localhost:9300
 
 build: 
   version: "@project.version@"
