+++ spectre/local-timer/src/main/java/guru/bonacci/spectre/localtimer/LocalTimerServer.java
@@ -6,7 +6,6 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.stream.annotation.EnableBinding;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.PropertySource;
 import org.springframework.web.client.RestTemplate;
 
 import guru.bonacci.spectre.spectreshared.events.SpectreEventChannels;
@@ -16,11 +15,9 @@ import guru.bonacci.spectre.spectreshared.persistence.PersistenceConfig;
 @EnableDiscoveryClient
 @EnableBinding(SpectreEventChannels.class)
 @Import(PersistenceConfig.class)
-@PropertySource("classpath:local-timer.properties")
 public class LocalTimerServer {
 	
 	public static void main(String[] args) {
-		System.setProperty("spring.config.name", "localtimer-server");
 		SpringApplication.run(LocalTimerServer.class, args);
 	}
