+++ spectre/local-timer/src/main/resources/bootstrap.yml
@@ -1,0 +1,7 @@
+spring:
+  application:
+    name: localtimer-service
+  cloud:
+    config:
+      uri: http://localhost:8888
+      fail-fast: true
