--- spectre/local-timer/src/main/resources/local-timer.properties
@@ -1,8 +1,0 @@
-# EVENTS
-# no content-type:json here. the message is sent from logstash where headers cannot be set
-spring.cloud.stream.default.binder=rabbit
-
-spring.cloud.stream.bindings.enrichment.group=localtimer
-
-# OTHER
-geo.name.username=leonardobonacci
