+++ spectre/local-timer/src/test/java/guru/bonacci/spectre/localtimer/services/LocalTimerServiceTests.java
@@ -1,4 +1,4 @@
-package guru.bonacci.spectre.money.services;
+package guru.bonacci.spectre.localtimer.services;
 
 
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -9,6 +9,9 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.NONE;
 
+import java.util.HashMap;
+import java.util.Map;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
@@ -17,34 +20,40 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.web.client.RestTemplate;
 
-import guru.bonacci.spectre.money.MoneyTestApp;
+import guru.bonacci.spectre.localtimer.LocalTimerTestApp;
+import guru.bonacci.spectre.localtimer.services.LocalTimerService;
 import guru.bonacci.spectre.spectreshared.persistence.Spec;
 import guru.bonacci.spectre.spectreshared.persistence.SpecRepository;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest(classes=MoneyTestApp.class, webEnvironment = NONE)
+@SpringBootTest(classes=LocalTimerTestApp.class, webEnvironment = NONE, properties = {
+	"geo.name.username=voldemort"		
+})
 @TestPropertySource("classpath:persistence-test.properties")
-public class MoneyServiceTest {
+public class LocalTimerServiceTests {
 
 	@Autowired
-	MoneyService service;
+	LocalTimerService service;
 	
 	@MockBean
 	SpecRepository repo;
 
 	@MockBean
-	MoneyCache cache;
+	RestTemplate rest;
 
 	@Test
 	public void shouldAddData() throws Exception {
 		Spec spec = new Spec();
 		spec.id = "ID";
-		spec.geoip.country_code2 = "NZ";
+		spec.geoip.latitude = 1.1;
+		spec.geoip.longitude = 2.2;
 		when(repo.findOne(spec.id)).thenReturn(spec);
 
-		String enrichmentData = "very little";
-		doReturn(enrichmentData).when(cache).get("NZ");
+		Map<String,Object> enrichmentData = new HashMap<>();
+		enrichmentData.put("a", "is not b");
+		doReturn(enrichmentData).when(rest).getForObject("http://api.geonames.org/timezoneJSON?lat=1.1&lng=2.2&username=voldemort", Map.class);
 
 		service.enrich(spec.id);
 
@@ -53,7 +62,7 @@ public class MoneyServiceTest {
 		ArgumentCaptor<Spec> arg3 = ArgumentCaptor.forClass(Spec.class);
 		verify(repo).addData(arg1.capture(), arg2.capture(), arg3.capture());
 
-		assertThat(arg1.getValue(), is(equalTo("income")));
+		assertThat(arg1.getValue(), is(equalTo("localtimer")));
 		assertThat(arg2.getValue(), is(equalTo(enrichmentData)));
 		assertThat(arg3.getValue(), is(equalTo(spec)));
 	}
