+++ spectre/money/src/main/resources/bootstrap.yml
