+++ spectre/money/src/main/resources/money-service.yml
@@ -1,11 +1,11 @@
 # Spring properties
 spring: 
   application: 
-    name: localtimer-service
+    name: money-service
 
 # HTTP Server
 server: 
-  port: 6666
+  port: 11111
 
 eureka: 
   client: 
@@ -27,5 +27,5 @@ logging:
     guru: 
       bonacci: 
         spectre: 
-          localtimer: INFO
-  file: C:\\logs\\spectre\\localtimer.log
+          money: INFO
+  file: C:\\logs\\spectre\\money.log
