--- spectre/money/src/main/resources/money.properties
@@ -1,5 +1,0 @@
-# EVENTS
-# no content-type:json here. the message is sent from logstash where headers cannot be set
-spring.cloud.stream.default.binder=rabbit
-
-spring.cloud.stream.bindings.enrichment.group=money
