+++ spectre/money/src/test/java/guru/bonacci/spectre/money/MoneyIntegrationTests.java
@@ -1,4 +1,4 @@
-package guru.bonacci.spectre.weather;
+package guru.bonacci.spectre.money;
 
 
 import static java.util.Collections.singletonMap;
@@ -9,8 +9,6 @@ import static org.mockito.Mockito.doReturn;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 import static org.springframework.messaging.MessageHeaders.CONTENT_TYPE;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.UUID;
 
 import org.junit.After;
@@ -37,28 +35,26 @@ import org.springframework.messaging.MessageHeaders;
 import org.springframework.messaging.support.MessageBuilder;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.web.client.RestTemplate;
 
-import guru.bonacci.spectre.weather.services.WeatherSpec;
-import guru.bonacci.spectre.weather.services.WeatherSpecRepository;
+import guru.bonacci.spectre.money.services.MoneyCache;
+import guru.bonacci.spectre.money.services.MoneySpec;
+import guru.bonacci.spectre.money.services.MoneySpecRepository;
 import guru.bonacci.spectre.spectreshared.events.SpectreEventChannels;
 import guru.bonacci.spectre.spectreshared.persistence.PersistenceTestConfig;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = RANDOM_PORT, properties = {
-	"openweathermap.apikey=1234567890"		
-})
+@SpringBootTest(webEnvironment = RANDOM_PORT)
 @TestPropertySource("classpath:persistence-test.properties")
-public class WeatherIntegrationTest {
+public class MoneyIntegrationTests {
 
 	@Autowired
 	BinderAwareChannelResolver resolver;
 
 	@Autowired 
-	WeatherSpecRepository repo;
+	MoneySpecRepository repo;
 
 	@MockBean
-	RestTemplate rest;
+	MoneyCache cache;
 
 	String uuid;
 
@@ -66,10 +62,9 @@ public class WeatherIntegrationTest {
 	public void init() {
 		uuid = UUID.randomUUID().toString();
 		
-		WeatherSpec spec = new WeatherSpec();
+		MoneySpec spec = new MoneySpec();
 		spec.id = uuid;
-		spec.geoip.latitude=1.1;
-		spec.geoip.longitude=2.2;
+		spec.geoip.country_code2="NZ";
 		repo.save(spec);
 	}
 
@@ -82,15 +77,14 @@ public class WeatherIntegrationTest {
 
 	@Test
 	public void shouldEventuallyAddData() {
-		Map<String,Object> enrichmentData = new HashMap<>();
-		enrichmentData.put("a", "is not b");
-		doReturn(enrichmentData).when(rest).getForObject("http://api.openweathermap.org/data/2.5/weather?lat=1.1&lon=2.2&appid=1234567890", Map.class);
+		String enrichmentData = "aaaaa lot";
+		doReturn(enrichmentData).when(cache).get("NZ");
 		
 		String body = "{\"content\":\"" + uuid + "\"}";
 		sendMessage(body, SpectreEventChannels.ENRICHMENT, "application/json");
 
-		WeatherSpec persisted = repo.findOne(uuid);
-		assertThat(persisted.weather.get("a"), is(equalTo("is not b")));
+		MoneySpec persisted = repo.findOne(uuid);
+		assertThat(persisted.income, is(equalTo("aaaaa lot")));
 	}
 
 	private void sendMessage(String body, String target, Object contentType) {
@@ -105,15 +99,15 @@ public class WeatherIntegrationTest {
 
 	@SpringBootApplication
 	@ComponentScan(excludeFilters = @Filter(type = FilterType.ASSIGNABLE_TYPE, 
-											value = {WeatherServer.class}))
+											value = {MoneyServer.class}))
 	@EnableBinding(SpectreEventChannels.class)
 	@IntegrationComponentScan
 	@EnableElasticsearchRepositories
 	@Import(PersistenceTestConfig.class)
-	static class WeatherIntegrationTestApp {
+	static class MoneyIntegrationTestApp {
 
 		public static void main(String[] args) {
-			SpringApplication.run(WeatherIntegrationTestApp.class, args);
+			SpringApplication.run(MoneyIntegrationTestApp.class, args);
 		}
 	}
 }
