+++ spectre/money/src/test/java/guru/bonacci/spectre/money/services/MoneyServiceTests.java
@@ -1,4 +1,4 @@
-package guru.bonacci.spectre.weather.services;
+package guru.bonacci.spectre.money.services;
 
 
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -9,9 +9,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.NONE;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
@@ -20,39 +17,34 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.web.client.RestTemplate;
 
-import guru.bonacci.spectre.weather.WeatherTestApp;
+import guru.bonacci.spectre.money.MoneyTestApp;
 import guru.bonacci.spectre.spectreshared.persistence.Spec;
 import guru.bonacci.spectre.spectreshared.persistence.SpecRepository;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest(classes=WeatherTestApp.class, webEnvironment = NONE, properties = {
-	"openweathermap.apikey=1234567890"		
-})
+@SpringBootTest(classes=MoneyTestApp.class, webEnvironment = NONE)
 @TestPropertySource("classpath:persistence-test.properties")
-public class WeatherServiceTest {
+public class MoneyServiceTests {
 
 	@Autowired
-	WeatherService service;
+	MoneyService service;
 	
 	@MockBean
 	SpecRepository repo;
 
 	@MockBean
-	RestTemplate rest;
+	MoneyCache cache;
 
 	@Test
 	public void shouldAddData() throws Exception {
 		Spec spec = new Spec();
 		spec.id = "ID";
-		spec.geoip.latitude = 1.1;
-		spec.geoip.longitude = 2.2;
+		spec.geoip.country_code2 = "NZ";
 		when(repo.findOne(spec.id)).thenReturn(spec);
 
-		Map<String,Object> enrichmentData = new HashMap<>();
-		enrichmentData.put("a", "is not b");
-		doReturn(enrichmentData).when(rest).getForObject("http://api.openweathermap.org/data/2.5/weather?lat=1.1&lon=2.2&appid=1234567890", Map.class);
+		String enrichmentData = "very little";
+		doReturn(enrichmentData).when(cache).get("NZ");
 
 		service.enrich(spec.id);
 
@@ -61,7 +53,7 @@ public class WeatherServiceTest {
 		ArgumentCaptor<Spec> arg3 = ArgumentCaptor.forClass(Spec.class);
 		verify(repo).addData(arg1.capture(), arg2.capture(), arg3.capture());
 
-		assertThat(arg1.getValue(), is(equalTo("weather")));
+		assertThat(arg1.getValue(), is(equalTo("income")));
 		assertThat(arg2.getValue(), is(equalTo(enrichmentData)));
 		assertThat(arg3.getValue(), is(equalTo(spec)));
 	}
