+++ spectre/sentiment/src/main/java/guru/bonacci/spectre/sentiment/SentimentServer.java
@@ -8,7 +8,6 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.stream.annotation.EnableBinding;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.PropertySource;
 
 import edu.stanford.nlp.pipeline.StanfordCoreNLP;
 import guru.bonacci.spectre.spectreshared.events.SpectreEventChannels;
@@ -18,11 +17,9 @@ import guru.bonacci.spectre.spectreshared.persistence.PersistenceConfig;
 @EnableDiscoveryClient
 @EnableBinding(SpectreEventChannels.class)
 @Import(PersistenceConfig.class)
-@PropertySource("classpath:sentiment.properties")
 public class SentimentServer {
 
 	public static void main(String[] args) {
-		System.setProperty("spring.config.name", "sentiment-server");
 		SpringApplication.run(SentimentServer.class, args);
 	}
