+++ spectre/sentiment/src/main/resources/bootstrap.yml
@@ -1,0 +1,7 @@
+spring:
+  application:
+    name: sentiment-service
+  cloud:
+    config:
+      uri: http://localhost:8888
+      fail-fast: true
