--- spectre/sentiment/src/main/resources/sentiment-server.yml
@@ -1,31 +1,0 @@
-# Spring properties
-spring: 
-  application: 
-    name: sentiment-service
-
-# HTTP Server
-server: 
-  port: 7777
-
-eureka: 
-  client: 
-    serviceUrl: 
-      defaultZone: http://localhost:1111/eureka/
-    instance: 
-      leaseRenewalIntervalInSeconds: 5
-
-#maven info
-build: 
-  version: "@project.version@"
-
-#logging
-logging: 
-  level: 
-    root: WARN
-    org: 
-      springframework: INFO
-    guru: 
-      bonacci: 
-        spectre: 
-          sentiment: INFO
-  file: C:\\logs\\spectre\\sentiment.log
