+++ spectre/sentiment/src/test/java/guru/bonacci/spectre/sentiment/SentimentIntegrationTests.java
@@ -1,16 +1,13 @@
-package guru.bonacci.spectre.localtimer;
+package guru.bonacci.spectre.sentiment;
 
 
 import static java.util.Collections.singletonMap;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
-import static org.mockito.Mockito.doReturn;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 import static org.springframework.messaging.MessageHeaders.CONTENT_TYPE;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.UUID;
 
 import org.junit.After;
@@ -21,7 +18,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.cloud.stream.annotation.EnableBinding;
 import org.springframework.cloud.stream.binding.BinderAwareChannelResolver;
 import org.springframework.context.annotation.Bean;
@@ -37,39 +33,32 @@ import org.springframework.messaging.MessageHeaders;
 import org.springframework.messaging.support.MessageBuilder;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.web.client.RestTemplate;
 
-import guru.bonacci.spectre.localtimer.services.LocalTimerSpec;
-import guru.bonacci.spectre.localtimer.services.LocalTimerSpecRepository;
+import guru.bonacci.spectre.sentiment.services.SentimentSpec;
+import guru.bonacci.spectre.sentiment.services.SentimentSpecRepository;
 import guru.bonacci.spectre.spectreshared.events.SpectreEventChannels;
 import guru.bonacci.spectre.spectreshared.persistence.PersistenceTestConfig;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = RANDOM_PORT, properties = {
-	"geo.name.username=voldemort"		
-})
+@SpringBootTest(webEnvironment = RANDOM_PORT)
 @TestPropertySource("classpath:persistence-test.properties")
-public class LocalTimerIntegrationTest {
+public class SentimentIntegrationTests {
 
 	@Autowired
 	BinderAwareChannelResolver resolver;
 
-	@Autowired 
-	LocalTimerSpecRepository repo;
-
-	@MockBean
-	RestTemplate rest;
-
 	String uuid;
+
+	@Autowired 
+	SentimentSpecRepository repo;
 
 	@Before
 	public void init() {
 		uuid = UUID.randomUUID().toString();
 		
-		LocalTimerSpec spec = new LocalTimerSpec();
+		SentimentSpec spec = new SentimentSpec();
 		spec.id = uuid;
-		spec.geoip.latitude=1.1;
-		spec.geoip.longitude=2.2;
+		spec.message = "what a wonderful wonderful life";
 		repo.save(spec);
 	}
 
@@ -82,15 +71,11 @@ public class LocalTimerIntegrationTest {
 
 	@Test
 	public void shouldEventuallyAddData() {
-		Map<String,Object> enrichmentData = new HashMap<>();
-		enrichmentData.put("a", "is not b");
-		doReturn(enrichmentData).when(rest).getForObject("http://api.geonames.org/timezoneJSON?lat=1.1&lng=2.2&username=voldemort", Map.class);
-		
 		String body = "{\"content\":\"" + uuid + "\"}";
-		sendMessage(body, SpectreEventChannels.ENRICHMENT, "application/json");
+		sendMessage(body, SpectreEventChannels.ENRICHMENT,"application/json");
 
-		LocalTimerSpec persisted = repo.findOne(uuid);
-		assertThat(persisted.localtimer.get("a"), is(equalTo("is not b")));
+		SentimentSpec persisted = repo.findOne(uuid);
+		assertThat(persisted.message, is(equalTo("what a wonderful wonderful life")));
 	}
 
 	private void sendMessage(String body, String target, Object contentType) {
@@ -102,18 +87,18 @@ public class LocalTimerIntegrationTest {
 	public MessageChannel routerChannel() {
 		return new DirectChannel();
 	}
-
+	
 	@SpringBootApplication
 	@ComponentScan(excludeFilters = @Filter(type = FilterType.ASSIGNABLE_TYPE, 
-											value = {LocalTimerServer.class}))
+											value = {SentimentServer.class}))
 	@EnableBinding(SpectreEventChannels.class)
 	@IntegrationComponentScan
 	@EnableElasticsearchRepositories
 	@Import(PersistenceTestConfig.class)
-	static class LocalTimerIntegrationTestApp {
+	static class SentimentIntegrationTestApp {
 
 		public static void main(String[] args) {
-			SpringApplication.run(LocalTimerIntegrationTestApp.class, args);
+			SpringApplication.run(SentimentIntegrationTestApp.class, args);
 		}
 	}
 }
