+++ spectre/spectre-shared/src/main/java/guru/bonacci/spectre/spectreshared/persistence/PersistenceConfig.java
@@ -2,12 +2,10 @@ package guru.bonacci.spectre.spectreshared.persistence;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
 import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories;
 
 @Configuration
 @ComponentScan
 @EnableElasticsearchRepositories
-@PropertySource("classpath:persistence.properties")
 public class PersistenceConfig {
 }
