--- spectre/spectre-shared/src/main/resources/persistence.properties
@@ -1,3 +1,0 @@
-# ES
-spring.data.elasticsearch.cluster-name=delphi
-spring.data.elasticsearch.cluster-nodes=localhost:9300
