+++ spectre/weather/src/main/java/guru/bonacci/spectre/weather/WeatherServer.java
@@ -6,7 +6,6 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.stream.annotation.EnableBinding;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.PropertySource;
 import org.springframework.web.client.RestTemplate;
 
 import guru.bonacci.spectre.spectreshared.events.SpectreEventChannels;
@@ -16,11 +15,9 @@ import guru.bonacci.spectre.spectreshared.persistence.PersistenceConfig;
 @EnableDiscoveryClient
 @EnableBinding(SpectreEventChannels.class)
 @Import(PersistenceConfig.class)
-@PropertySource("classpath:weather.properties")
 public class WeatherServer {
 	
 	public static void main(String[] args) {
-		System.setProperty("spring.config.name", "weather-server");
 		SpringApplication.run(WeatherServer.class, args);
 	}
