--- spectre/weather/src/main/resources/weather.properties
@@ -1,8 +1,0 @@
-# EVENTS
-# no content-type:json here. the message is sent from logstash where headers cannot be set
-spring.cloud.stream.default.binder=rabbit
-
-spring.cloud.stream.bindings.enrichment.group=weather
-
-# OTHER
-openweathermap.apikey=9c721267a69a906658d20d673c73b1a9
