+++ utilities/src/test/java/guru/bonacci/oogway/utils/MyListUtilsTests.java
@@ -17,7 +17,7 @@ import org.junit.runner.RunWith;
 import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
-public class MyListUtilsTest {
+public class MyListUtilsTests {
 
 	@Test
 	public void shouldReturnDifferentElements() {
