+++ web/src/test/java/guru/bonacci/oogway/web/services/WebControllerTests.java
@@ -23,7 +23,7 @@ import guru.bonacci.oogway.web.WebTestApp;
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = WebTestApp.class, webEnvironment = RANDOM_PORT)
 @AutoConfigureMockMvc
-public class WebControllerTest {
+public class WebControllerTests {
 
 	@Autowired
 	MockMvc mvc;
